% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{add_move_stats}
\alias{add_move_stats}
\title{Add move to stats df}
\usage{
add_move_stats(
  move,
  B_N,
  B_S,
  r_N,
  c_N,
  d_N,
  r_S,
  c_S,
  d_S,
  depth,
  stats_df = stats_df
)
}
\arguments{
\item{B_N}{board n}

\item{B_S}{board s}

\item{stats_df}{}
}
\value{

}
\description{
Add move to stats df
}
