% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_components.R
\name{capture_rule}
\alias{capture_rule}
\title{capture_rule}
\usage{
capture_rule(player, B_N, B_S, end_pit_r, end_pit_c)
}
\arguments{
\item{player}{Player making the move, \code{"S"} for south or \code{"N"} for north}

\item{B_N}{Board of player N}

\item{B_S}{Board of player S}

\item{end_pit_r, end_pit_c}{Row and column of the pit where the sowing ended}
}
\value{
\code{list(captured = cap, N = B_N, S= B_S, stones_won = st_won)} Boolean indicating if capture took place, boards of player N and S
  after the capture, number of stones captured. If capture did not take place
  it returns \code{FALSE}
}
\description{
Decide if capture is activated at the end of the move
}
