% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{check_and_move}
\alias{check_and_move}
\title{Check if starting pit is allowed and perform move.}
\usage{
check_and_move(player, B_N, B_S, pit_r, pit_c, direction, print_int = F)
}
\arguments{
\item{print_int}{}
}
\description{
Check if starting pit is allowed and perform move.
}
