% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{check_if_end_game}
\alias{check_if_end_game}
\title{Check if game should end or continue given boards of players.}
\usage{
check_if_end_game(B_N, B_S, print_int = FALSE)
}
\arguments{
\item{print_int}{}
}
\description{
Check if game should end or continue given boards of players.
}
