% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{choose_random_move}
\alias{choose_random_move}
\title{Choose random but possible move.}
\usage{
choose_random_move(board)
}
\arguments{
\item{board}{Board from which we want to draw move}
}
\value{
a valid move
}
\description{
Choose random but possible move.
}
