% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_components.R
\name{decode_board}
\alias{decode_board}
\title{Decode board from string to matrix}
\usage{
decode_board(board_encoded)
}
\arguments{
\item{board_encoded}{}
}
\value{

}
\description{
Decode board from string to matrix
}
