% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_components.R
\name{encode_board}
\alias{encode_board}
\title{Encode board from matrix to string}
\usage{
encode_board(board)
}
\arguments{
\item{board}{a matrix representing a HusBao board}
}
\value{
a string representing the matrix
}
\description{
Encode board from matrix to string
}
