% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{human_player}
\alias{human_player}
\title{Human player}
\usage{
human_player(name, state)
}
\arguments{
\item{name}{name of the player "N" for north and "S" for south}

\item{state}{state of the game}
}
\value{
move
}
\description{
take move from human player and parse it to vector
}
