% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{init_stats_df}
\alias{init_stats_df}
\title{Initialize df with game statistics}
\usage{
init_stats_df()
}
\value{
an empty data frame with columns: move, move_N, move_S, stones_N,
stones_S, depth, B_N, B_S
}
\description{
Initialize df with game statistics
}
