% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{Mancala}
\alias{Mancala}
\title{An R6 class to represent a mancala game}
\description{
An R6 class to represent a mancala game

An R6 class to represent a mancala game
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{players}}{a list of functions taking the state of the game and
returning the move.}

\item{\code{state}}{contains the boards of both players}

\item{\code{stats_df}}{is a dataframe containing statistics of the played game}

\item{\code{active_player}}{the player currently taking the turn}

\item{\code{check_game_end}}{Checker of whether game should end. Function taking state of the game as an input and returning a logical.}

\item{\code{make_move}}{Mechanics of move. Function that takes the state of the game, the player's identifier and a their move and returns state of the game}

\item{\code{init_state}}{Initial state of the game. Function that returns the initial state of the game.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Mancala$new()}}
\item \href{#method-reset_state}{\code{Mancala$reset_state()}}
\item \href{#method-play_game}{\code{Mancala$play_game()}}
\item \href{#method-clone}{\code{Mancala$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mancala$new(
  player_N = mancalaR::human_player,
  player_S = mancalaR::human_player,
  check_game_end = mancalaR::check_if_end_game,
  make_move = mancalaR::check_and_move,
  init_state = mancalaR::init_state,
  starting_state = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{player_N}}{Mechanics of player N. Function that takes the state of the game as an input and returns a move as a numeric vector.}

\item{\code{player_S}}{Mechanics of player S. Function that takes the state of the game as an input and returns a move as a numeric vector}

\item{\code{check_game_end}}{Checker of whether game should end. Function taking state of the game as an input and returning a logical.}

\item{\code{make_move}}{Mechanics of move. Function that takes the state of the game, the player's identifier and a their move and returns state of the game}

\item{\code{init_state}}{Initial state of the game. Function that returns the initial state of the game.}

\item{\code{starting_state}}{optional. Starting state. A state from which we want to start the game instead of the initial state.

reset state}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset_state"></a>}}
\if{latex}{\out{\hypertarget{method-reset_state}{}}}
\subsection{Method \code{reset_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mancala$reset_state()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
assigns initial state to the state field
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-play_game"></a>}}
\if{latex}{\out{\hypertarget{method-play_game}{}}}
\subsection{Method \code{play_game()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mancala$play_game(print_int = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{print_int}}{logical. choose whether to show verbose debugging stuff}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mancala$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
