% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{parse_move_str}
\alias{parse_move_str}
\title{Parse move to easily handeled information}
\usage{
parse_move_str(r, c, d)
}
\arguments{
\item{r}{row of pit}

\item{c}{col of pit}

\item{d}{direction}
}
\value{
string in the form of "rcd"
}
\description{
Parse move to easily handeled information
}
