% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{random_vs_random_mode}
\alias{random_vs_random_mode}
\title{Two bots}
\usage{
random_vs_random_mode(B_N, B_S, print_int = FALSE)
}
\arguments{
\item{print_int}{}
}
\value{

}
\description{
Each chooses random move out of possible moves until the game ends.
}
