% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{simulate_mancala}
\alias{simulate_mancala}
\title{Simulate a number of mancala games}
\usage{
simulate_mancala(mode = "random_vs_random", gameplay = play_game, n = 20)
}
\arguments{
\item{mode}{chosen mode, available the same as in \code{play_mancala}}

\item{gameplay}{function that plays a single game}

\item{n}{number of times that the game should be played}
}
\value{
dataframe consisting of statistics of \code{n} games played
}
\description{
Simulate a number of mancala games
}
