% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{single_sow}
\alias{single_sow}
\title{single_sow}
\usage{
single_sow(player, B_N, B_S, pit_r, pit_c, direction)
}
\arguments{
\item{player}{Player making the move, S for south or N for north}

\item{B_N}{Board of player N}

\item{B_S}{Board of player S}

\item{pit_r}{Row of the starting pit}

\item{pit_c}{Column of the starting pit}

\item{direction}{Direction of sowing, either 1 for clockwise of 0 for
counter-clockwise}
}
\value{
Boards of player N and S, ending pit coordinates and boolean
  indicating whether we've ended in an empty pit
}
\description{
Make a single sowing move, without "compound" sowing
}
