% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{two_player_mode}
\alias{two_player_mode}
\title{Two player console mode}
\usage{
two_player_mode(B_N, B_S, print_int = FALSE)
}
\arguments{
\item{B_N}{board of player N}

\item{B_S}{board of player S}

\item{print_int}{logical. choose whether to print verbose debugging communicates}
}
\value{
dataframe with game statistics
}
\description{
Two player console mode
}
