% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{add_move_stats}
\alias{add_move_stats}
\title{Add move to stats df}
\usage{
add_move_stats(
  move,
  B_N,
  B_S,
  r_N,
  c_N,
  d_N,
  r_S,
  c_S,
  d_S,
  depth,
  stats_df = stats_df
)
}
\arguments{
\item{move}{number of move}

\item{B_N}{board of player N}

\item{B_S}{board of player S}

\item{r_N}{row of N move}

\item{c_N}{column of N move}

\item{d_N}{direction of N move}

\item{r_S}{row of S move}

\item{c_S}{direction of S move}

\item{d_S}{column of S move}

\item{depth}{length of the move (number of single sows)}

\item{stats_df}{the df to which the row should be added}
}
\description{
Add move to stats df
}
