% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_components.R
\name{check_if_move_possible}
\alias{check_if_move_possible}
\title{Check if provided move is possible}
\usage{
check_if_move_possible(pit_r, pit_c, board)
}
\arguments{
\item{pit_r}{row of the chosen pit}

\item{pit_c}{column of the chosen pit}

\item{board}{board of the player making the move}
}
\description{
Check if provided move is possible
}
