% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{make_move}
\alias{make_move}
\title{make_move}
\usage{
make_move(
  player,
  B_N,
  B_S,
  pit_r,
  pit_c,
  direction,
  should_end = FALSE,
  can_capture = FALSE,
  print_int = FALSE,
  move_length = 0
)
}
\arguments{
\item{move_length}{}
}
\value{
state of the game (boards and move length)
}
\description{
make_move
}
