% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{Mancala}
\alias{Mancala}
\title{An R6 class to represent a mancala game}
\description{
An R6 class to represent a mancala game

An R6 class to represent a mancala game
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{player_N, player_S}}{Functions taking the state of the game and
returning the move.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{player_N, player_S}}{Functions taking the state of the game and
returning the move.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Mancala$new()}}
\item \href{#method-reset_state}{\code{Mancala$reset_state()}}
\item \href{#method-play_game}{\code{Mancala$play_game()}}
\item \href{#method-clone}{\code{Mancala$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mancala$new(
  player_N = mancalaR::human_player,
  player_S = mancalaR::human_player,
  capture_rule = mancalaR::capture_rule,
  check_game_end = mancalaR::check_if_end_game,
  make_move = mancalaR::check_and_move,
  init_state = mancalaR::init_state,
  starting_state = NA
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset_state"></a>}}
\if{latex}{\out{\hypertarget{method-reset_state}{}}}
\subsection{Method \code{reset_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mancala$reset_state()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-play_game"></a>}}
\if{latex}{\out{\hypertarget{method-play_game}{}}}
\subsection{Method \code{play_game()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mancala$play_game(print_int = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mancala$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
