% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{two_player_mode}
\alias{two_player_mode}
\title{Two player console mode}
\usage{
two_player_mode(B_N, B_S, print_int = FALSE)
}
\arguments{
\item{print_int}{}
}
\value{

}
\description{
Two player console mode
}
\examples{



}
